+++ Controls/_columns/ColumnsControl.wml
@@ -2,9 +2,14 @@
                            moveMarkerOnScrollPaging="{{false}}"
                            plainItemsContainer="{{false}}"
                            _keyDownHandler="{{_keyDownHandler}}"
+                           beforeMountCallback="{{_beforeMountCallback}}"
                            name="baseControl">
     <!--прокидываем itemTemplate т.к. иначе мы не понимаем, что нужно перестраивать partial, потому что изменения только в замыкании функции-->
     <Controls._columns.ColumnsInnerView name="innerView">
-        <ws:partial template="{{_options.viewName}}" itemTemplate="{{_options.itemTemplate}}"/>
+        <ws:partial template="{{_options.viewName}}"
+                    itemTemplate="{{_options.itemTemplate}}"
+                    columnsCount="{{_columnsCount}}"
+                    spacing="{{_spacing}}"
+                    on:resize="_resizeHandler()"/>
     </Controls._columns.ColumnsInnerView>
 </Controls.list:BaseControl>
