+++ Controls/_columns/render/resources/For.wml
@@ -17,6 +17,7 @@
               on:longtap="_onItemLongTap(item)"
               on:mouseenter="_onItemMouseEnter(item)"
               on:mousedown="_onItemMouseDown(item)"
+              on:mouseup="_onItemMouseUp(item)"
               on:mousemove="_onItemMouseMove(item)"
               on:mouseleave="_onItemMouseLeave(item)"
               on:wheel="_onItemWheel(item)"
