+++ Controls/_list/BaseControl.ts
@@ -5968,6 +5968,11 @@ const BaseControl = Control.extend(/** @lends Controls/_list/BaseControl.prototy
         }
     },
 
+    // Уйдет когда будем наследоваться от baseControl
+    getItemsContainer() {
+        return this._getItemsContainer();
+    },
+
     _itemsContainerReadyHandler(_: SyntheticEvent<Event>, itemsContainerGetter: Function): void {
         this._getItemsContainer = itemsContainerGetter;
         this._viewReady = true;
